+++ js/extensions/to-string.js
@@ -1,0 +1,14 @@
+define([
+    'jquery'
+], function($) {
+
+/**
+ * Return the markup of the element in the collection as a string.
+ *
+ * @returns {String}
+ */
+$.fn.toString = function() {
+    return this.prop('outerHTML');
+};
+
+});